fix!: remove lavalink v1
