+++ DisCatSharp.LavalinkV1/Entities/LavalinkUpdates.cs
@@ -26,7 +26,7 @@ using System;
 
 using Newtonsoft.Json;
 
-namespace DisCatSharp.Lavalink.Entities;
+namespace DisCatSharp.LavalinkV1.Entities;
 
 /// <summary>
 /// The lavalink state.
