+++ DisCatSharp.LavalinkV1/GlobalSuppressions.cs
