+++ DisCatSharp.LavalinkV1/LavalinkTrack.cs
@@ -29,7 +29,7 @@ using System.Runtime.Serialization;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace DisCatSharp.Lavalink;
+namespace DisCatSharp.LavalinkV1;
 
 /// <summary>
 /// Represents a lavalink track.
