+++ DisCatSharp.LavalinkV1/global.json
